// baggage-tracker-ui/src/app/enums/page.enum.ts
export enum Page {
	Login = 'login',
	Home = 'home',
	QrCodeScan = 'qr-code-scan',
	PostScan = 'post-scan',
	GenerateQrCode = 'generate-qr-code',
	ActiveFlights = 'active-flights',
	RegisterFlightManifest = 'register-flight-manifest',
	TrackBaggages = 'track-baggages',
	TrackBaggagesByFlight = 'track-baggages-by-flight',
	Help = 'help'
}